CSW-258: a more resilient django webapp deployment script

+++ setup.sh
@@ -15,13 +15,12 @@
 #     a. copies and configures the code for one of the 5 UCB deployments
 #     b. "npm builds" the needed js components
 #     c. if running on a UCB server (which is detected automatically), rsyncs the code to the runtime directory
-# 3. other maintainance functions: 'updatejs' just rebuilds the webpack, 'disable' or 'enable' individual webapps
+# 3. other maintainance functions: 'disable' or 'enable' individual webapps
 #
 
 # exit on errors...
 # TODO: uncomment this someday when the script really can be expected to run to completion without errors
 # set -e
-set -x
 
 COMMAND=$1
 # the second parameter can stand for several different things!
@@ -32,7 +31,6 @@ DEPLOYMENT=$2
 # nb: version is optional. if not present, current repo, with or without changes is used...
 VERSION="$3"
 
-CURRDIR=`pwd`
 CONFIGDIR=~/cspace-webapps-ucb
 
 PYTHON=python3
@@ -40,32 +38,51 @@ PYTHON=python3
 function buildjs()
 {
     # TODO: fix this hack to make the small amount of js work for all the webapps
-    export TENANT="$TENANT"
-    perl -i -pe 's/..\/..\/suggest/\/$ENV{TENANT}\/suggest/' client_modules/js/PublicSearch.js
+    if [[ "$OSTYPE" == "linux-gnu" ]]; then
+        export TENANT="$TENANT"
+        perl -i -pe 's/..\/..\/suggest/\/$ENV{TENANT}\/suggest/' client_modules/js/PublicSearch.js
+    fi
 
     npm install
     npm build
     ./node_modules/.bin/webpack
-    # disable this for now, until we address the errors it detects
+    # disable eslint for now, until we address the errors it detects
     #./node_modules/.bin/eslint client_modules/js/app.js
 }
 
 function deploy()
 {
     buildjs $1
-    # if this is running on a dev or prod system (evidenced by the presence of web-accessible
-    # deployment directories, i.e. /var/www/*), then copy the needed files there
-    # nb: the config directory is not overwritten!
-    YYYYMMDD=`date +%Y%m%d`
-    RUNDIR=~/${YYYYMMDD}/$1
-    mkdir -p ${RUNDIR}
-    if [[ -e ${RUNDIR} ]]; then
+    # if this is running on an RTL server, the runtime directory is ~/M/YYYYMMDD
+    # (where M is the museum and YYYYMMDD is today's date)
+    # if not Linux, e.g. Darwin (= development), configure everything in the current directory ...
+    if [[ "$OSTYPE" == "linux-gnu" ]]; then
+        YYYYMMDD=`date +%Y%m%d`
+        RUNDIR=~/$1/${YYYYMMDD}
+        if [[ -e ${RUNDIR} ]]; then
+            echo "Cowardly refusal to overwrite existing runtime directory ${RUNDIR}"
+            echo "Remove/rename ${RUNDIR}, then try again."
+            exit 1
+        fi
+        mkdir -p ${RUNDIR}
         # copy the built files to the runtime directory, but leave the config files as they are
         rsync -av --delete --exclude node_modules --exclude .git --exclude .gitignore --exclude config . ${RUNDIR}
-	cd ${RUNDIR}
+
+        # copy the most recent backup of the config files into this new runtime directory
+        # nb: presumes that these are the right ones! any changes to configuration needed
+        # for this release will need to be applied (by hand, presumably) in advance
+        # of relinking.
+        cp -r ~/backup/$1/config ${RUNDIR}
+
+        cd ${RUNDIR}
     fi
 
-    cp -r ~/backup/$1/config ${RUNDIR}
+    echo "*************************************************************************************************"
+    echo "The configured CSpace system is:"
+    grep 'hostname' config/main.cfg
+    echo "*************************************************************************************************"
+
+    # now we can go ahead and complete the configuration
     $PYTHON manage.py migrate --noinput
     $PYTHON manage.py loaddata fixtures/*.json
     # get rid of the existing static_root to force django to rebuild it from scratch
@@ -79,9 +96,11 @@ function check_version()
     if [[ $VERSION != "" ]]; then
         if [[ $(git status -s) ]]; then
             echo
-            echo 'fyi, uncommitted changes or untracked files were found and version $VERSION was specified.'
-            echo 'initial deployments of a particular version must be from a "clean" branch.'
-            echo 'cowardly refusal to proceed; please clean up and try again...'
+            echo 'FYI, uncommitted changes or untracked files were found and version $VERSION was specified.'
+            echo
+            echo 'Initial deployments of a particular version must be from a "clean" branch.'
+            echo
+            echo 'Cowardly refusal to proceed; please clean up and try again...'
             echo
             echo 'e.g. something like the following'
             echo
@@ -95,12 +114,13 @@ function check_version()
         git checkout ${VERSION}
 
     else
-        echo "no version specified; deploying code as is, but with specified customizations."
+        echo
+        echo "No version specified; deploying code as is, but with specified customizations."
     fi
 }
 
 if [ $# -lt 2 -a "$1" != 'show' ]; then
-    echo "Usage: $0 <enable|disable|deploy|updatejs|configure|show> <TENANT|CONFIGURATION|WEBAPP> (VERSION)"
+    echo "Usage: $0 <enable|disable|deploy|configure|show> <TENANT|CONFIGURATION|WEBAPP> (VERSION)"
     echo
     echo "where: TENANT = 'default' or the name of a deployable tenant"
     echo "       CONFIGURATION = <pycharm|dev|prod>"
@@ -116,7 +136,7 @@ if [ $# -lt 2 -a "$1" != 'show' ]; then
 fi
 
 if [[ ! -e manage.py ]]; then
-    echo "no manage.py found. this script must be run in the django project directory"
+    echo "No manage.py found. This script must be run in the django project directory"
     echo
     exit 1
 fi
@@ -124,13 +144,13 @@ fi
 if [[ "${COMMAND}" = "disable" ]]; then
     perl -i -pe "s/('${WEBAPP}')/# \1/" cspace_django_site/installed_apps.py
     perl -i -pe "s/(path)/# \1/ if /${WEBAPP}/" cspace_django_site/urls.py
-    echo "disabled ${WEBAPP}"
+    echo "Disabled ${WEBAPP}"
 elif [[ "${COMMAND}" = "enable" ]]; then
     perl -i -pe "s/# *('${WEBAPP}')/\1/" cspace_django_site/installed_apps.py
     perl -i -pe "s/# *(path)/\1/ if /${WEBAPP}/" cspace_django_site/urls.py
     python manage.py migrate --noinput
     python manage.py collectstatic --noinput
-    echo "enabled ${WEBAPP}"
+    echo "Enabled ${WEBAPP}"
 elif [[ "${COMMAND}" = "show" ]]; then
     echo
     echo "Installed apps:"
@@ -139,7 +159,7 @@ elif [[ "${COMMAND}" = "show" ]]; then
     echo
 elif [[ "${COMMAND}" = "configure" ]]; then
     if [[ ! -f "cspace_django_site/extra_${DEPLOYMENT}.py" ]]; then
-        echo "can't configure '${DEPLOYMENT}': use 'pycharm', 'dev', or 'prod'"
+        echo "Can't configure '${DEPLOYMENT}': use 'pycharm', 'dev', or 'prod'"
         echo
         exit
     fi
@@ -157,19 +177,19 @@ elif [[ "${COMMAND}" = "configure" ]]; then
     echo
 elif [[ "${COMMAND}" = "deploy" ]]; then
     if [[ ! -d "${CONFIGDIR}" ]]; then
-        echo "the repo containing the configuration files (${CONFIGDIR}) does not exist"
-        echo "please either create it (e.g. by cloning it from github)"
+        echo "The repo containing the configuration files (${CONFIGDIR}) does not exist"
+        echo "Please either create it (e.g. by cloning it from github)"
         echo "or edit this script to set the correct path"
         echo
-        exit
+        exit 1
     fi
 
     # checkout indicated version, if any...
     check_version
 
     if [[ ! -d "${CONFIGDIR}/${TENANT}" ]]; then
-        echo "can't deploy tenant ${TENANT}: ${CONFIGDIR}/${TENANT} does not exist"
-        echo
+        echo "Can't deploy tenant ${TENANT}: ${CONFIGDIR}/${TENANT} does not exist"
+        echo 1
         exit
     fi
 
@@ -177,11 +197,11 @@ elif [[ "${COMMAND}" = "deploy" ]]; then
     # we don't check the version for the 'example' repo...
     # cd ${CONFIGDIR}
     # check_version
-    # cd ${CURRDIR}
 
-    rm config/*
-    rm fixtures/*
+    rm -f config/*
+    rm -f fixtures/*
 
+    # use 'default' configuration for this tenant from github, initially.
     cp ${CONFIGDIR}/${TENANT}/config/* config
     cp ${CONFIGDIR}/${TENANT}/fixtures/* fixtures
     # note that in some cases, this cp will overwrite customized files in the underlying contributed apps
@@ -190,10 +210,6 @@ elif [[ "${COMMAND}" = "deploy" ]]; then
     cp ${CONFIGDIR}/${TENANT}/project_urls.py cspace_django_site/urls.py
     cp ${CONFIGDIR}/${TENANT}/project_apps.py cspace_django_site/installed_apps.py
     cp client_modules/static_assets/cspace_django_site/images/header-logo-${TENANT}.png client_modules/static_assets/cspace_django_site/images/header-logo.png
-    echo "*************************************************************************************************"
-    echo "configured system is:"
-    grep 'hostname' config/main.cfg
-    echo "*************************************************************************************************"
     # just to be sure, we start over with the database...
     rm -f db.sqlite3
 
@@ -205,9 +221,7 @@ elif [[ "${COMMAND}" = "deploy" ]]; then
     echo
     echo "*************************************************************************************************"
     echo "Don't forget to check cspace_django_site/main.cfg if necessary and the rest of the"
-    echo "configuration files in config/ (these are .cfg and .csv files)"
-    echo
-    echo "please restart apache to pick up changes"
+    echo "configuration files in config/ (these are .cfg, .json, and .csv files mostly)"
     echo "*************************************************************************************************"
     echo
 else
